IOException -> UncheckedIOException
This is the library which targets running in Java Web application.
Web frameworks like JAX-RS and Spring, provided ways to process
Exception globally.
And checked exceptions, in some cases, force users to handle low-level
error handling and follow the 'throws' interface.
I don't think it is not *public interface* in this library at least.